Refactor asset trust toggler
